использование глобального логгера.
