+++ Diarisation/diarazer.py
@@ -3,12 +3,13 @@ import datetime
 from config import settings
 from Diarisation.do_diarize import load_and_preprocess_audio
 from utils.pre_start_init import posted_and_downloaded_audio
-from utils.do_logging import logger
-
 from collections import defaultdict
+import logging
+logger = logging.getLogger(__name__)
 
 if settings.CAN_DIAR:
     from Diarisation import diarizer
+
 
 async def do_diarizing(
         file_id:str,
