+++ VoiceActivityDetector/__init__.py
@@ -1,8 +1,8 @@
 from config import settings
 from utils.pre_start_init import paths
-from utils.do_logging import logger
 import requests
-
+import logging
+logger = logging.getLogger(__name__)
 
 if not paths.get("vad_model_path").exists():
     logger.info("Модель silero_vad.onnx отсутствует. Предпринимаем попытку скачать её.")
