+++ routes/legacy/post_by_url.py
@@ -3,12 +3,13 @@ import asyncio
 import os
 from fastapi import APIRouter
 from utils.pre_start_init import posted_and_downloaded_audio
-from utils.do_logging import logger
 from utils.get_audio_file import getting_audiofile, open_default_audiofile
 from models.fast_api_models import SyncASRRequest, BaseResponse
 
 from fastapi import Depends
 from Recognizer import get_recognizer, Recognizer
+
+from Punctuation import get_punctuator, SbertPuncCaseOnnx
 
 from Recognizer.engine.file_recognition import process_file
 from threading import Lock
@@ -22,7 +23,8 @@ audio_lock = Lock()
 
 @router.post("/post_one_step_req", response_model=BaseResponse)
 async def post(params: SyncASRRequest,
-                recognizer: Recognizer = Depends(get_recognizer)
+                recognizer: Recognizer = Depends(get_recognizer),
+                punctuator: SbertPuncCaseOnnx = Depends(get_punctuator)
 ) -> BaseResponse:
     """
     На вход ждёт str(HttpUrl) - прямую ссылку на скачивание файла 'mp3', 'wav' или 'ogg'.\n
@@ -60,7 +62,11 @@ async def post(params: SyncASRRequest,
 
     try:
         # Запускаем обработку в потоке
-        result_dict = await asyncio.to_thread(process_file, posted_and_downloaded_audio[post_id], params, recognizer)
+        result_dict = await asyncio.to_thread(process_file,
+                                              tmp_path =posted_and_downloaded_audio[post_id],
+                                              params=params,
+                                              recognizer=recognizer,
+                                              punctuator=punctuator)
         result = BaseResponse(**result_dict)
     except Exception as e:
         error_description = f"Ошибка обработки в process_file - {e}"
