+++ routes/v1/post_by_file_FORM.py
@@ -1,11 +1,14 @@
 from io import BytesIO
 import asyncio
-from utils.do_logging import logger
 from config import settings
 from fastapi import APIRouter, Depends, File, Form, UploadFile
 from models.fast_api_models import PostFileRequest, V1ASRResponse, ASRData, RawData, SentencedData, DiarizedData
 from Recognizer.engine.file_recognition import process_file
 from Recognizer import get_recognizer, Recognizer
+from Punctuation import get_punctuator, SbertPuncCaseOnnx
+
+import logging
+logger = logging.getLogger(__name__)
 
 router = APIRouter()
 # Функция для извлечения параметров из FormData
@@ -41,7 +44,9 @@ def get_file_request(
 async def async_receive_file(
     file: UploadFile = File(description="Аудиофайл для обработки"),
     params: PostFileRequest = Depends(get_file_request),
-    recognizer: Recognizer = Depends(get_recognizer)
+    recognizer: Recognizer = Depends(get_recognizer),
+    punctuator: SbertPuncCaseOnnx = Depends(get_punctuator)
+
 ) -> V1ASRResponse:
     try:
         buffer = BytesIO(await file.read())
@@ -57,7 +62,11 @@ async def async_receive_file(
     else:
         logger.info(f"Получен и сохранён файл {file.filename}")
         try:
-            result_dict = await asyncio.to_thread(process_file, buffer, params, recognizer)
+            result_dict = await asyncio.to_thread(process_file,
+                                                  tmp_path=buffer,
+                                                  params=params,
+                                                  recognizer=recognizer,
+                                                  punctuator=punctuator)
             return V1ASRResponse(
                 success=result_dict.get('success', True),
                 error_description=result_dict.get('error_description'),
