+++ utils/chunk_doing.py
@@ -1,7 +1,6 @@
 from config import settings
 import numpy as np
 from pydub import AudioSegment
-from utils.do_logging import logger
 from utils.bytes_to_samples_audio import get_np_array_samples_float32
 
 from utils.pre_start_init import (audio_overlap,
@@ -11,6 +10,8 @@ from utils.pre_start_init import (audio_overlap,
 
 from VoiceActivityDetector import vad
 from utils.resamppling import async_resample_audiosegment
+import logging
+logger = logging.getLogger(__name__)
 
 
 async def find_last_speech_position(socket_id, is_last_chunk):
